Improve variable names by avoiding generic 'data' variable name.
The 'geocode' function was particularly hard to follow because it uses
both 'this.data' and also 'data'. It's not particularly clear what
what the difference is from the name, and it doesn't help that 'data'
can be singular or plural.

I replaced `this.data' with `this.records`, which exactly matches the
name used in the REST API. Accordingly, a single record is referred as
`record`.

With names that mirror the API and with clear singular and plural
nounds, it's easier to follow what's going on.